CL1: Add interfaces and stubs for granular memory tracking.

+++ services/resource_coordinator/public/cpp/memory_instrumentation/detailed_metrics_delegate.h
@@ -28,14 +28,14 @@ namespace memory_instrumentation {
 // Optimized structure for smaps metrics.
 struct COMPONENT_EXPORT(RESOURCE_COORDINATOR_PUBLIC_MEMORY_INSTRUMENTATION)
     SmapsMetrics {
-  uint64_t rss_kb = 0;
-  uint64_t pss_kb = 0;
-  uint64_t private_dirty_kb = 0;
-  uint64_t private_clean_kb = 0;
-  uint64_t shared_dirty_kb = 0;
-  uint64_t shared_clean_kb = 0;
-  uint64_t swap_kb = 0;
-  uint64_t swap_pss_kb = 0;
+  uint32_t rss_kb = 0;
+  uint32_t pss_kb = 0;
+  uint32_t private_dirty_kb = 0;
+  uint32_t private_clean_kb = 0;
+  uint32_t shared_dirty_kb = 0;
+  uint32_t shared_clean_kb = 0;
+  uint32_t swap_kb = 0;
+  uint32_t swap_pss_kb = 0;
 };
 
 // Interface for project-specific detailed memory metrics collection.
@@ -51,7 +51,7 @@ class COMPONENT_EXPORT(RESOURCE_COORDINATOR_PUBLIC_MEMORY_INSTRUMENTATION)
 
   // Swaps the internal results map with an empty one and returns it via |stats|.
   virtual void GetAndResetStats(
-      base::flat_map<std::string, uint64_t>* stats) = 0;
+      base::flat_map<std::string, uint32_t>* stats) = 0;
 
   virtual base::WeakPtr<DetailedMetricsDelegate> GetWeakPtr() = 0;
 };
@@ -59,4 +59,3 @@ class COMPONENT_EXPORT(RESOURCE_COORDINATOR_PUBLIC_MEMORY_INSTRUMENTATION)
 }  // namespace memory_instrumentation
 
 #endif  // SERVICES_RESOURCE_COORDINATOR_PUBLIC_CPP_MEMORY_INSTRUMENTATION_DETAILED_METRICS_DELEGATE_H_
-
